+++ src/ClassField.hs
@@ -6,16 +6,16 @@ import Data.Matrix
 import Types
 import Const
 
--- | Поле есть матрица цветов
-type Field = Matrix Color
-
 -- | Создание изначального поля
 createField :: Int -> Int -> Field
 createField x y = matrix x y (initFunction x y)
 -- | Начальное заполнение фона
 initFunction :: Int -> Int -> ((Int,Int)->Color)
-initFunction x y =
-  ( \_ -> makeColor 0.1 0.5 0.2 1.0)
+initFunction _ _ =
+  (\_ -> makeColorI 34 139 34 255)
+  -- ^ веселья ради можно поставить что-то ещё
+  -- но цвет лесной зелени приятен глазу, как листва деревьев в летнем саду
+  -- >>> ( \_ -> makeColor 0.13 0.54 0.13 1.0)
 
 -- | обновление поля - добавление в него серий бросков, числом от дельты времени
 updateField :: StdGen -> viewPoint -> Float -> Field -> Field
@@ -23,10 +23,8 @@ updateField gR _ dt field =
   generator
   gR
   field
-  3 --(floor (dt*numCast))
-  -- 2 в центре пр-уг
-  -- 3 ниже центра пр-уг
-  0
+  (floor (dt*numCast))
+  0 -- ^ просто для того, чтобы нумерация была удобной
 -- | генератор нового поля
 generator :: StdGen -> Field -> Int -> Int -> Field
 --generator g f m n | n < m  = rty (iter (f,(busPoint g n)) 0) m (n+1)
@@ -37,10 +35,6 @@ generator g f m n | n < m  = rty ( temp (f,(busPoint g n)) 0) m (n+1)
     pack newC@(cast, _) = ((plot cast f), newC)
 generator _ f _ _  = f
 
--- | Точка и цвет в карте градиентов [0,1)
-type Cast = (Vec, Double)
--- | Бросок с привязанным генератором
-type CastGen = ((Vec, Double),StdGen)
 -- | Iterator for loop inner_iter
 -- new Field, new PRNG
 -- броски одной точки
@@ -65,9 +59,29 @@ busPoint g i = (busPointList !! i, g)
 -- | Cast Infinite List
 busPointList :: [Cast]
 busPointList = 
-  [(point,colC) | point <- biUnitTiling]
+  [(point, colC) | point <- biUnitTiling]
   where
     colC = 0.5
+-- | соседи одной точки, расположенные в центрах окружающих квадратов
+-- порядок обхода - по контуру
+-- полиморф только из-за повторного использования в Gloss simulate
+-- можно восстановить getNeigbours :: Double -> Vec -> [Vec]
+getNeigbours::Num a =>  a->(a,a)->[(a,a)]
+getNeigbours dl (x,y) = [v11,v12,v22,v21]
+  where
+    v11 = (x+dl,y+dl)
+    v12 = (x+dl,y-dl)
+    v21 = (x-dl,y+dl)
+    v22 = (x-dl,y-dl)
+-- | список соседей 1 порядка
+nthNeigbours :: Int -> [Vec]
+nthNeigbours n | n>0 = concat $ map (getNeigbours dl) (nthNeigbours (n-1))
+  where
+    dl = 2 ** (- fromIntegral n)
+nthNeigbours _ = [(0,0)]
+-- | бесконечный список соседей
+biUnitTiling :: [Vec]
+biUnitTiling = concat [ nthNeigbours i | i <- [0,1..]]
 
 -- | Размещение точки в поле
 plot :: Cast -> Field -> Field
